Flash-resident program image with final-id save (format v2) (#31)
Coordinated persisted-format break: S1 envelope v2 (u32 payload length) and
payload v5, with the profile hash extended to cover image-layout caps so a
profile can never mis-mount another's flash image. Old v1/payload-4 saves are
rejected cleanly and boot fresh.

Saved code becomes immutable image bytes the mount points at directly: on ESP32
the frothy partition's active slot is memory-mapped (esp_partition_mmap) and the
pointer-based VM executes straight from flash; on host the committed payload
blob backs it. Mount validates and indexes only -- it copies and patches no code,
text, or name bytes. All id/ref rewriting moves to a save-time finalize that
assigns dense final image ids deps-before-dependents and rewrites every ref kind
(PUSH_CODE_ID and PUSH_OBJECT_ID operands, bind/slot value refs, record shape and
field refs, event bodies, and tagged words in cells). The image code arena is
gone from RAM -- the first real RAM payback.

Mount validation rejects records claiming ids in the immutable base partition,
enforces dense above-base ids, and rejects deps-after-dependent order; save
rejects cyclic code refs. Restore is validate-then-swap: the running mapping is
kept until a candidate fully validates, a failed restore drops all transient
image code back to the base boundary so no entry outlives its mapping, and the
A/B corrupt-newer fallback recovers the older good image. The existing tier
side-table is preserved, with a mounted-base tier stamp added because overlay
state is intentionally cleared after remount.

Round-trip identity and resolve-equivalence tests exercise all six ref kinds;
new regressions cover below-base rejection, cyclic-ref rejection, event-body
survival across remount, failed-restore candidate drop, and corrupt-newer
fallback. Verified on the DevKit: full-language smoke, a persisted event
executing from the mmapped flash image with no fault, save/reboot/restore
round-trip, and corrupt-newer fallback to the older good image. The diff was
independently reviewed.

+++ src/crc.c
@@ -1,7 +1,7 @@
 #include "crc.h"
 
-uint32_t fr_crc32_update(uint32_t crc, const uint8_t *bytes, uint16_t length) {
-  for (uint16_t i = 0; i < length; i++) {
+uint32_t fr_crc32_update(uint32_t crc, const uint8_t *bytes, uint32_t length) {
+  for (uint32_t i = 0; i < length; i++) {
     crc ^= bytes[i];
     for (uint8_t bit = 0; bit < 8; bit++) {
       uint32_t mask = 0u - (crc & 1u);
@@ -11,6 +11,6 @@ uint32_t fr_crc32_update(uint32_t crc, const uint8_t *bytes, uint16_t length) {
   return crc;
 }
 
-uint32_t fr_crc32(const uint8_t *bytes, uint16_t length) {
+uint32_t fr_crc32(const uint8_t *bytes, uint32_t length) {
   return ~fr_crc32_update(0xffffffffu, bytes, length);
 }
